pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brother_ql"
version = "1.0.0"
description = "Python 3.13+ library for Brother QL label printers - Modernized by Luxardo Labs"
readme = "README.md"
requires-python = ">=3.13"
license = "GPL-3.0-or-later"
authors = [
    {name = "Philipp Klaus", email = "philipp.l.klaus@web.de"},
    {name = "Luxardo Labs"},
]
maintainers = [
    {name = "Luxardo Labs"},
]
keywords = [
    "Brother", "QL-500", "QL-550", "QL-560", "QL-570", "QL-700", 
    "QL-710W", "QL-720NW", "QL-800", "QL-810W", "QL-820NWB", 
    "QL-1050", "QL-1060N", "label", "printer", "driver",
    "Python3.13", "modernized", "type-hints", "enums"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "Topic :: Printing",
    "Typing :: Typed",
]

dependencies = [
    "click>=8.0",
    "packbits>=0.6",
    "Pillow>=10.0",
    "pyusb>=1.2.0",
    "attrs>=22.0",
]

[project.urls]
"Original Repository" = "https://github.com/pklaus/brother_ql"
"Modernized Fork" = "https://github.com/luxardolabs/brother_ql"
Homepage = "https://github.com/luxardolabs/brother_ql"
Documentation = "https://github.com/luxardolabs/brother_ql"
Repository = "https://github.com/luxardolabs/brother_ql"
Issues = "https://github.com/luxardolabs/brother_ql/issues"


[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-Pillow",
]

[tool.setuptools.packages.find]
include = ["brother_ql*"]

[tool.setuptools.package-data]
brother_ql = ["config/*.json"]

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long - handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]